graph search working, but evaluation of paths is too mixed in. about to switch to split out path evaluation from path discovery again, which is what was here previously.
Graph searching (ie., path discovery) happens iteratively, and messily.
Therefore, if path discovery and evaluation are mixed (as they are in this commit),
then when evaluating a path we can't know what came 'before' the node being
evaluated - as that's an attirbute of the path, not the discovery algorithim.

Therefore, we NEED this to be two distinct phases.

The difference between the Hipster and OrientDb approaches will be that
with Hipster we can still keep path evaluation relatively dynamic, by retaining
a context of known / knowable facts, and using that information to derive if
paths are traversable when constructing the path