refactor(schema-registration): when schemas are registered, include some form of package information.
This allows the UI to show the schema packages that are registered (more useful than just a list of file names),
and show more meaningful summarized feedback as schema registration evolves over time.

Pausing on this work, as it's not going to be part of our 5-min/1-hour journey.  However, here's where I'm up to:

 * Server-side building blocks are in place - added submitSchemaPackages() and routed most schema registration through that.
 * New server endpoint in schemaServer for exposing a list of schema summaries.
 * No UI work started

 TODO :
  * Schemas that are registered through config (eg., file schemas, git schemas, etc) need to be tested to ensure their schema summaries are returned
  * A large number of failing tests, though these appear to be because mocks aren't being fired - likely because schema registration methods have changed