refactor(build): replace spring-boot-maven plugin for building query-service, as it breaks protobuf generation
We're using Wire (https://github.com/square/wire) for protobuf generation.
At runtime, this uses okio (https://github.com/square/okio) generate artifacts in order to generate a profobuf schema - and it treats
the current classpath as a zip.  This is reasonable, because normally the jar is a valid zip.
However, spring boot rewires this to make an executable, fat jar.  That breaks okio, and therefore wire.

Ideally, there'd be a workaround, but I haven't found one.
See discussions here:
https://github.com/square/okio/issues/1094
https://github.com/square/wire/discussions/2209

So, we're reverting to a classic jar build using maven appassembler