fix(ui/results-tree): fix results tree showing empty when the results change.
Previously, the results tree used an array, with items pushed to it.  However, the component requires onPush change detection, so this broke, and also meant trees couldn't be used for streaming results.

the result tree / ObjectView component now uses an observable.

Also, removed the usage of type info in the object-view.component.ts, as it broke whenever using anonymous types.