fix(typed-instance-parser): when a field defines a default value, use the actual value if available.
Historically, we haven't actually ever read the value of a field if a default is defined, we just return the default.

Pipelines and Projections mean that a default value could actually be discoverable.

This introduces a first-pass implementation that will use the actual value if present, otherwise defer to the default.  There are stack overflow considerations here, so this implementation disables accessor evaluation while determining if the actual value is readable.

This could be improved later.