fix(core): address perf issues - anon types and query compilation caching
This removes anonymous types being stored in the schema.
Previously, types were appended to the schema, which lead to the type cache being constantly mutated (initially in an non threadsafe way).  Rather than fix the threading issues, this PR removes the persisted state between queries.

There are a number of side effects, which have been resolved - but mainly around type construction, where we used to rely on anon types being in the schema, but now we rely on type data present in fields.