feat(core): support multiple variable scopes when projecting
When projecting, you sometimes need to hoist multiple variables into the scope. This is to make them available at the top level of a projection, without declaring them directly in the output.

For example

In the below example, we want to operate on a single item from an array of values, as well as the parent object.

However, we need both the objects in scope:

```
find { Film } as (Film, first(Actor[])) -> {  // two values declared in scope
            title : Title // title comes from Film
            starring : ActorName // ActorName comes from Actor
         }
```

ORB-225