fix(core/querying): query responses now returned as object or array depending on content.
Previously, we were sending all responses as an Array, even if the underlying result only contained a single item, and was intended as an object.

The root cause is that we were returning things as a Flow<> or Flux<> depending on the endpoint.

We now return as either Flux<> or Mono<> (abstracted as CorePublisher<>). For this to work, API endpoints cannot be suspend functions. (Spring cannot handle a suspend function returning a Mono or Flux - partly because it doesn't really make sense to mix frameworks like that!)

Also, the API for dispatched queries has changed from Flux<> to Mono<>, which better reflects what's actually happening today. We currently don't support streamed results from remote queries, so everything is a Mono<> anyway.

If we migrate to streaming results from remote executors (which we should), the remote sources need a way of signalling "This response has a single item / This response has a list".