feat: Add schema-diagram to data source creation step
This chunk of work was fundamentally about adding the schema-diagram
alongside the type-viewer so the user could see the changes they make
to the type mappings have a correlation to the connections surfaced.
This was mostly achieved by combining the schema with the partialSchema
in the schema-member-type-explorer component and passing that to the
schema-diagram.

The schema-diagram now has hoverable states which allows the user to
visualise the connections between the nodes and edges more easily.

Also:
- Move the point at which the site overflows it's content, ensuring the
  search in the header always remains on screen
- Re-style onboarding container for a more consistent look and feel
- Add a project-import/data-source route
- Prevent navigating to types from schema-diagram when in editable mode
- Correct color for Type node and edges in schema-diagram
- Change styling of connector nodes in schema-diagram
- Tighten up spacing in Elk layout
- Attempt to surface correct Types/Operations for Models/Attribute in
  schema-diagram but still some work to be done to make this 100% correct
- Remove angular-resize-event dependency, replacing with a service (one
  step closer to being able to npm i without -force) and ensure that the
  triggering of the refresh of the schema-diagram works (mostly)
- a bunch of little styling tweaks across pretty much the entire app

Closes: ORB-242