feat: Rework data source into connections page
Allow the existing data source import flow to be accessed from a "Data
source" sidebar route (replacing the existing "Connection" one).

The connections are now surfaced using a tui-tree, which allows quicker
navigation from each connection. It also allows surfacing the details
related to each connection, and making clearer the hierarchy of a
connection to the project it's associated with.

Also:
- lots of components made standalone, allowing the new routing for the
  Data source page to work correctly
- revamped the connection-status component, bringing it into alignment
  with the badge treat from the project-list
- introduction of data-source-tree component
- add 'requiredAllAuthorities' AuthGuard property to require multiple
  privileges for an AuthGuard to pass
- add hover state to as-split divider

Closes: ORB-186