feat: Add form validation to data source imports
Previously the user could hit the "Configure" button with gay abandon prior to filling out the salient details on the form that are required. They can't do that now after this.

Also fixed a few things that were just broken, namely protobuf file selection and made some tweaks to the Kafka form removing Material and making the service/operation inputs nicer.

The JSON schema import (which is currently disabled) also was pretty broken, so that's now working (for when we enable it again).

And finally added a bit of validation logic to the project imports as well (didn't want them to feel left out)

Closes: ORB-223