bug: Improve streaming query performance
With the addition of some Rx buffering, and allowing the user to pause the stream this should hopefully tide us over for a bit. 

Added maxQueryRecordCount to AppConfig to configure how many records can be shown at one time.

When a query is streaming, the results are inverted with the last one appearing at first, and any records after the maxQueryRecordCount amount are culled from the UI. 

Quite a few styling tweaks to the Query cards in the query-history list

Closes: ORB-337