feat: Tabbed query editing
This introduces the ability for the user to have multiple tabs editable
in the query-editor/panel. Any queries that have been run will persist
in the results area, and are persisted for the session.

The vast majority of the business logic in the query-editor component
has been moved into the query-editor.state, essentiall mapping a tab to
a query-editor.state class. These instances are managed by the
query-editor-store.service

The tabs and persistence of them to localStorage is managed by the
query-panel-store.service

Closes: ORB-365