feat: Add endpoint info to Endpoint route
Allow endpoints to be copied as URL and cURL and standardise into a
component for reuse in both the query-editor and the Endpoints route.

Have added a rather generic sounding dropdown component (currently
resides in the query-panel/query-editor/query-editor-toolbar folder, but
can be hoisted if this is appropriate) and used that to genericise the
copy functionality required in the query-editor-toolbar and other places

Restrict endpoints from having paths that already exist

Reorganised the query-editor-toolbar related components into their own
folder as well.

Closes: ORB-385, ORB-406