perf(hazelcast): use our existing non-hzc cache as a near cache for hazelcast
This introduces a significant performance boost.
Previously, we were hitting Hazelcast on each request for data from a remote query.
However, the graph query engine requests from data sources frequently, and repeatidly.
The intent was that this query behaviour would always be paired with a cache, so that
we weren't incurring network traffic. However, the original design of the hazelcast cache
uses a remote cache, so all requests incur network I/O. Therefore, even if a single
request is fast, the aggregate becomes very very slow.

This fixes the issue, by adding a local cache - which is our existing operation cache.