feat(mysql): added support for MySQL
Known issues:

DDL
When generating a table, if the column type is string, and the column is used in an index, then the size must be set. We default to VARCHAR(255), which cannot be changed currently
MySQL doesn't support create index if not exists, so we've had to fall back to create index. This means that the creation may fail if the index already exists
DML
MySQL (and therefore JOOQ) doesn't support UPDATE ... RETURNING syntaxes. This means on an upsert, we are unable to find the id of the created entry (see related jOOQ issue)

ORB-408