fix: Only show project health badge when unhealthy
Hopefully removes confusion around when the project has unhealthy data
sources as well.

Fixed issue on dashboard where the the data-sources were stuck in a
loading loop when a datasource wasn't defined correctly. Now surface the
same information as the data-source/problems route when this occurs.

Update websocket.service connect method with a genericised return

Closes: ORB-416