chore: Apply correct cache strategy for web resources
This ensures that even with caching enabled in DevTools, making changes
to files and then reloading the app, the files changed are reloaded and
everything that hasn't changed loads from the cache.

Fixes the main issue which was that even with caching enabled, static
resources were always being reloaded everytime they were accessed.