test(project-import): Add e2e tests
This is the first swing at getting e2e tests working for the git and
local disk project import functionality.

Various fixes to the UI forms in this area were made, including
validation and making the results of the service call correctly typed.

There's still work to be done in the git-config and file-config
components as these use a ngForm directive to handle the form state which
restricts what can be done with validation messages. Ideally this gets
overhauled to use a FormGroup/FormControl approach which will interface
better with Taiga as well.

The Playwright version was bumped which allows the tests to be run in a
interactive UI (npm run start-ui) which greatly speeds up development of
the tests and allows creating locators within the UI.

Lastly an auth flow has been added for the tests to allow users to enter
their credentials into a .env file which Playwright will then use to
authenticate before the tests are run. This approach will need to be
augmented once we reinstate the e2e tests in CI land.

Refs: ORB-460