fix: login form not shown when session expired
Fourth attempt at fixing this! The actual fix was a combination of the MC IdP allowing /silent-refresh.html, along with re-enabling the silentRefreshRedirectUri prop in the AuthConfig. Now when the session is terminated from outside of Flow, a silent_refresh_timeout legitimately occurs, so we can proceed with the initLoginFlow call.

Closes: ORB-492