feat: Improve onboarding of model designer
Add help text and disable panels so as to guides users through the steps involved in creating a model that maps to the source they've uploaded or written.

Added 8Mb upload limit and improved UX of file upload process. 

Group new types separate to existing types in the type-autocomplete comp

Closes: ORB-507