fix: Result set shows inconsistent numbers
This stops the maxQueryRecordCount affecting records being culled in the various components, instead we now pause the display of anymore records once the number of records shown exceeds maxQueryRecordCount and alert the user to this. The user can then choose to continue showing more results if they want their UI to blow up.

Also made maxQueryRecordCount configurable on the server

Closes: ORB-536