feat: Allow adding/removing schema diagram nodes
Reworked the node design to use a single row for the header, integrated
the icon for Services into the header (much to Marty's chagrin) and
tweaked the animations applied to the nodes and their edges.
Added a schema-multi-select to allow the user to add/remove nodes from
the diagram.

Also:
- Upgraded react-flow to the latest version
- Fixed storybook schema-diagram stories

Closes: ORB-579