fix(playground): navigation from empty route now works
Previously navigating from the root route would cause the
VoyagerAppComponent to instantiate itself again which breaks the diagram

This paved the way for the correct operation of the clear button, which
now resets the route and clears the schema/query out as well.

Also added routerActiveLink to the examples menu and added some active
styling so the user knows which example they're looking at.