feat: allow copying markdown of schema diagram nodes and positions
To facilitate this, introduced a new layout mode in the schema diagram,
'NONE'. This will allow the user to lay out nodes, without Elk trying to
send things flying all around the screen.

This new mode is automatically selected if there's position information
in the schema diagram from the markdown wrapper.

Closes: ORB-580