fix(query-visualizer): improve rendering of query plans by deferring to operation invokers
Previously, we treated all calls as HTTP operations in the query visualizer.

We now allow invokers to implement a plan() method, which allows returning an RemoteCall showing what WOULD'VE happened if the call was a real call.

A default implementation is provided, so we don't have to update all the invokers in one go.

ORB-874