feat(streams): fix error reporting in streams - if a stream dies it is now...
feat(streams): fix error reporting in streams - if a stream dies it is now accurately reported in the ui

The challenge in building this was that there's two distinct concepts in play here:

Stream State (eg: Enabled / Disabled) - which indicates the overall
desired state of the stream.

Job state - which is the state of the underlying jet jobs actually executing
the stream.

A Stream can have multiple jobs over the course of it's lifespan, especially
if the jobs fail, or are restarted.

This is now more accurately modelled

ORB-868