fix(core): factbag searches return null if the target type is an attribute of a null instance
For example:

find { Film } as {
   filmAward : AwardTitle //comes from film
   // find the star of the film (CastMember), and return their AwardTitle
   starAward = first(CastMember[], (IsStar) -> IsStar == true) as AwardTitle
}

In the above example, if first(CastMember[]) returns null, then the StarAward must also be null.
However, if we don't yeild a null here, then we'll search elsewhere in higher scopes, and find the
AwardTitle from the parent object - a Film.

Contributes to ORB-877, ORB-878, ORB-879