fix(graph-builder): add all known facts to graph at start, and search from a new 'start' node
This commit contains an important change.

Previously, when doing graph searches, we used to:
 * Build the graph
 * Iterate each known fact, using it as a start point, and search from there

If a fact returned a path, we'd use it.

However, this is incorrect, as if there are multiple facts, and `Fact1` has a shorter path than `Fact0`, it's not used, as we use the path yielded by `Fact0`.

So, this PR changes this:
 * We now add all facts to the graph
 * Starting facts are linked to a special "start node"
 * All searches go from the start node.

This has a few benefits:
 * Less graph building
 * Fewer graph searches
 * We'll now serve the best path across all known facts.