fix(kafka): when a kafka connection is changed or the schema changes,...
fix(kafka): when a kafka connection is changed or the schema changes, transparently reconnect the kafka subscription; disable kafka info message emission (see below for why)

Modifies the Kafka subscriptions to have two emitters:
 - a long-lived public emitter, which consumers attach to
  - an internal emitter, which attaches to Kafka instances and emits into the public emitter. This is destroyed and rebuilt whenever the schema changes

This fixes an issue where queries running against Kafka when a schema changes do not reflect the updated schema of the emitted models.

NOTE: This also removes the emission of info messages via the KafkaConsumerStatsFlowBuilder. A critical bug is found in that class that causes the server to become unresponsive when a Kafka broker goes away. As this was a quick-hack for a customer, it's no longer being used, so have disabled the code and added a warning.