draft: Add high level stream error metrics
This adds low-cardinality reporting when errors occur on a stream.
This specifically attempts to address an issue where projections for mutations fail (because not all the inputs are available), and there's no visibility.

Unfortunately, it's difficult for us currently to associate this back to either the stream that's running or the originating message, but those details are available in the logs.