feat(map): Modify map { } behaviour to enable iteration of a collection to...
feat(map): Modify map { } behaviour to enable iteration of a collection to call single-input endpoint.

Users need to be able to handle scnearios like:

```
write operation saveOne( PersonName, ApiKey, PersonUpdate ) : PersonUpdate
```

Where the operation takes multiple inputs.
The query accepts an array:

```
@HttpOperation(method = "POST", url = "/api/q/submit-single-calls")
query submitCsv(@RequestBody body: StockReportCsv[]) {
...
```

And we want to iterate that array, and call the mutation, correclty handling projections (Input[] -> Input -> Output)

By using `map {}` we preserve the rules of `find {}`, which would want to treat `A[] -> B` as an aggregation, not an iteration.